Bundle preloads with esbuild
